Lift out the implementation of substitution
